requirements.txt
amqp==5.0.6
appdirs==1.4.3
asgiref==3.3.1
autopep8==1.5.5
billiard==3.6.4.0
CacheControl==0.12.6
celery==5.0.5
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
Django==3.1.7
django-3-jet==1.0.8
django-celery-beat==2.2.0
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-js-asset==1.2.2
django-modeltranslation==0.16.2
django-timezone-field==4.1.2
djangorestframework==3.12.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
kombu==5.0.2
lockfile==0.12.2
Markdown==3.3.4
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
Pillow==8.1.1
progress==1.5
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.6
python-crontab==2.5.1
python-dateutil==2.8.1
python-slugify==4.0.1
python3-openid==3.2.0
pytoml==0.1.21
pytz==2021.1
redis==3.5.3
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.25.8
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
